refactor(api): PR4 — split positions into api/positions.py + db/positions.py
Move position CRUD queries (db_create_position, db_get_positions,
db_close_position, db_update_position) to db/positions.py. _calc_pnl
(pure math helper) lives in db/positions.py and is re-exported by
api/positions.py to avoid circular imports.

Move position routes (/positions GET POST, /positions/{id} PUT DELETE,
/positions/{id}/close POST), service helpers (_calc_pnl,
check_position_stops, _write_position_event_log, update_positions_json)
to api/positions.py. check_position_stops notification path uses the
notifier module via lazy import (unchanged from original).

btc_api.py re-exports the DB functions and check_position_stops +
update_positions_json for legacy callers (scanner_loop runs
check_position_stops per symbol per cycle) until PR7.

Updated tests/test_api.py TestPositionsAPI.client fixture to reference
route functions from api.positions instead of btc_api (they no longer
live in btc_api module scope after the move).

Parity test (tests/test_api_positions_parity.py) backed by snapshot
baseline covers: GET all/open/closed, POST without auth (401),
POST with auth (200 + new position), PUT (set notes). Dynamic
entry_ts field is stripped before comparison.

1009 tests pass.
